feat: admin page api utils
관리자 페이지에서 사용될 api 유틸함수들
전체 유저 목록 리스트 조회 및
역할 변경 요청 api
